refactor: remove width prop and move to new directory
